src/main/java/payrollcasestudy/transactions/add/salesreceipt/AddSalesReceiptTransaction.java
package payrollcasestudy.transactions.add.salesreceipt;

import payrollcasestudy.Employee;
import payrollcasestudy.PayrollDatabase;
import payrollcasestudy.SalesReceipt;
import payrollcasestudy.paymentclassifiactions.CommissionedPaymentClassification;
import payrollcasestudy.paymentclassifiactions.PaymentClassification;
import payrollcasestudy.transactions.Transaction;

import java.util.HashMap;
import java.util.Map;

public class AddSalesReceiptTransaction implements Transaction{

    private int date;
    private double amount;
    private int employeeId;

    public AddSalesReceiptTransaction(int date, double amount, int employeeId) {
        this.date = date;
        this.amount = amount;
        this.employeeId = employeeId;
    }

    public void execute() {
        Employee employee = PayrollDatabase.globalPayrollDatabase.getEmployee(employeeId);
        if (employee != null){
            PaymentClassification paymentClassification = employee.getPaymentClassification();
            if (paymentClassification instanceof CommissionedPaymentClassification){
                CommissionedPaymentClassification commissionedPaymentClassification =
                        (CommissionedPaymentClassification) paymentClassification;
                commissionedPaymentClassification.addSalesReceipt(new SalesReceipt(date, amount));
            }
        }
    }
}
